add home 1-4
Find PI to the Nth Digit – Enter a number and have the program generate
PI up to that many decimal places. Keep a limit to how far the program
will go.

Fibonacci Sequence – Enter a number and have the program generate the
Fibonacci sequence to that number or to the Nth number.

Prime Factorization – Have the user enter a number and find all Prime
Factors (if there are any) and display them.

Next Prime Number – Have the program find prime numbers until the user
chooses to stop asking for the next one.